+++ gem/lib/ecs_rails.rb
@@ -12,6 +12,7 @@ require "ecs_rails/presence"
 require "ecs_rails/validations"
 require "ecs_rails/dsl"
 require "ecs_rails/querying"
+require "ecs_rails/preloading"
 require "ecs_rails/entity"
 require "ecs_rails/component"